refactor: split navigation logic into separate service files
