+++ src/shared/services/index.ts
@@ -1,0 +1,3 @@
+export * from './routes'
+export * from './navigation-list'
+export * from './navigation-permissions'
