+++ src/shared/services/navigation-list.ts
@@ -1,0 +1,31 @@
+import routes from "./routes";
+
+export const navigationList = [
+  {
+    name: 'content',
+    text: 'Контент',
+    children: [
+      {
+        name: 'job',
+        text: 'Работа',
+        children: [routes.vacancies, routes.candidates],
+      },
+      {
+        name: 'news',
+        text: 'Новости',
+        children: [routes.events],
+      },
+    ],
+  },
+  {
+    name: 'users',
+    text: 'Пользователи',
+    children: [
+      {
+        name: 'inner-users',
+        text: 'Внутренние пользователи',
+        children: [routes.clients, routes.partners],
+      },
+    ],
+  },
+]
