+++ src/shared/services/routes.ts
@@ -1,0 +1,33 @@
+const routes = {
+  vacancies: {
+    name: 'vacancies',
+    pathname: 'vacancies',
+    getLink: () => '/vacancies',
+    text: 'Вакансии',
+  },
+  candidates: {
+    name: 'candidates',
+    pathname: 'candidates',
+    getLink: () => '/candidates',
+    text: 'Кандидаты',
+  },
+  events: {
+    name: 'events',
+    pathname: 'events',
+    getLink: () => '/events',
+    text: 'События',
+  },
+  clients: {
+    name: 'clients',
+    pathname: 'clients',
+    getLink: () => '/clients',
+    text: 'Клиенты',
+  },
+  partners: {
+    name: 'partners',
+    pathname: 'partners',
+    getLink: () => '/partners',
+    text: 'Клиенты',
+  },
+}
+export default routes
